+++ mcu/attiny-hal/src/attiny84.rs
@@ -1,0 +1,5 @@
+pub mod adc {
+    pub use crate::periphals::adc::*;
+
+    // Fixme: Implement ADC for ATtiny84.
+}
