+++ mcu/attiny-hal/src/periphals/mod.rs
@@ -1,0 +1,3 @@
+// ATtiny2313 does not have ADC and will not compile with this module
+#[cfg(not(feature = "attiny2313"))]
+pub mod adc;
